week3/Task11/requirements.txt
openai==1.88.0
python-dotenv==1.1.0
mutagen==1.47.0
pydub==0.25.1
